Add SCSS styles and processing rules
Based on Phil Hawksworth's excellent advice[1] this sets up inline CSS which is pre-compiled
byt a gulp task that watches for scss file changes that then triggers and rebuild by eleventy
when the output file causes eleventy to rebuild the site.

SCSS purely used for ease as it's what I use and will get annoyed if I don't include it.

Could have tried plain CSS and CSS vars and so on, but time is short so sticking with
what I know for this iteration of the auction site.

1. https://www.hawksworx.com/blog/keeping-sass-simple-and-speedy-on-eleventy/