Add Netlify lambda function to email after bid made
We should email bid confirmations that allow the user to see their bid
info after the fact. Using SendGrid node package and the
submission-created.js event driven netlify function this should be
possible.

The item data was missing from the form, so that has been added for
email and associating the bids to an item, which wouldn't be possible
previously.