Add Asyncronous form submission to link to correct confirmation page
Th Netlify action was redirecting to the correct URL but showing the
wrong content. So by submitting with JavaScript and redirecting manually
means that the form will redirect correctly with JS enabled and work
fine without.

Next enhancment might be not leaving th page at all and using an overlay
or replace the form with a success message, so the user still has the
next and previous navigation available.